fix: 실수 값의 수익률이 나오지 않던 오류 수정

+++ src/main/java/lotto/LottoPrizeRecordingManager.java
@@ -33,7 +33,7 @@ public class LottoPrizeRecordingManager {
     public double getPrizeRate(long purchaseMoney) {
         setTotalPrizeMoney();
 
-        return totalPrizeMoney / purchaseMoney;
+        return (double)totalPrizeMoney / purchaseMoney;
     }
